requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
Django==4.2
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
kombu==5.2.4
numpy==1.24.3
pandas==2.0.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
